+++ Authoring/TestDiagnostics/StructField_StructInNamespaceTests.cs
@@ -1,0 +1,128 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace TestDiagnostics
+{
+
+    public sealed class SillyClass
+    {
+        public double Identity(double d)
+        {
+            return d;
+        }
+    }
+
+    public struct StructWithClass_Invalid
+    {
+        public SillyClass classField;
+    }
+
+    public struct StructWithDelegate_Invalid
+    {
+        public delegate int ADelegate(int x);
+    }
+
+    public struct StructWithEvent_Invalid
+    {
+        public event EventHandler<int> EventField;
+    }
+
+    public struct StructWithConstructor_Invalid
+    {
+        int X;
+        StructWithConstructor_Invalid(int x)
+        {
+            X = x;
+        }
+    }
+
+    public struct StructWithIndexer_Invalid
+    {
+        int[] arr;
+        int this[int i] => arr[i];
+    }
+
+    public struct StructWithMethods_Invalid
+    {
+        int foo(int x)
+        {
+            return x;
+        }
+    }
+    public struct StructWithConst_Invalid // really invalid ?
+    {
+        const int five = 5;
+    }
+    */
+
+    /*
+    public enum BasicEnum
+    { 
+        First = 0,
+        Second = 1
+    }
+
+    public struct Posn_Invalid 
+    {
+        BasicEnum enumField; 
+
+        public int x { get; }
+        public int y { get; }
+    }
+    */
+
+    /*
+    public struct StructWithPrivateField_Invalid
+    {
+        const int ci = 5;
+        private int x;
+    }
+    */
+
+    /*
+    public struct StructWithAllValidFields
+    {
+        bool boolean;
+        char character;
+        decimal dec;
+        double dbl;
+        float flt;
+        int i;
+        uint nat;
+        long lng;
+        ulong ulng;
+        short sh;
+        ushort us;
+        string str;
+    }
+    */
+
+    /*
+    public struct StructWithObjectField_Invalid // is this really invalid? 
+    {
+        object obj;
+
+    }
+    public struct StructWithDynamicField_Invalid // is this really invalid? 
+    {
+        dynamic dyn;
+    }
+
+    public struct StructWithByteField_Invalid // is this really invalid? 
+    {
+        byte b;
+    }
+
+    public struct StructWithWinRTStructField
+    {
+        public Matrix3x2 matrix;
+    }
+    */
+
+    class StructField_StructInClassTests
+    {
+    }
+}
